Fix broken error labels
The invalidText prop was deprecated and no longer works. With these
changes I am transitioning into using the helperText prop for errors,
but this requires some changes in the overall logic of error handling.

It would be better to have a formik wrapper to TextInput where this
error logic could be applied in a single place, but I didn't spend the
time here to do that. THese changes will make it easier to go towards
that direction.