Provide authorization code for GDPR calls
The GDPR API requires a single use token which allows it to know that
the user has allowed for their data to be accessed. This code is
created in the UI because consenting may require user actions.

These changes add the generation of this token into profile download
and delete flows.