Fix missing login button in build
The specificity of CSS rules changed between development build and
production build. This caused HDS to override some styles in
production, which caused the button to go missing.

In this commit I fixed this issue by adding specificity into the
selectors for our custom rules. I added a question into the HDS repo
in an attempt to uncover a better fix.