Remove bottom margin from expanding panel
With this change we don't have to synchronize margin values between
components. Instead, the components can be agnostic about how sparsley
they should be positioned within the page. Instead the page can, for
instance with the help of a grid, assing and control whitespace. This
way whitespace ends up being managed in fewer places, limiting the
maintenance effort.