Centralised queries and profile data
ProfileContext is a React Context provider, so any component can access the profile data from anywhere. It is not required to use other data storages or re-query data - unless it is being mutated. When data changes, it is changed everywhere. Data is loaded after user is authenticated. ProfileContext also tracks data loading and provides easy getName()-function.

Data queries are moved to hooks.ts. It is used for reading and updating/mutating data. It also updates the Apollo cache when mutation is complete on server side. Hooks.ts and mutationEditor.ts are used by UI-components to update data.

Due to a feature/bug in Apollo, data provided via Apollo's hooks change sometimes uncontrollably. Data is loaded like it is supposed and Apollo's 'loading' status can change to false. Like it should.
But another update from Apollo's hook may follow immediately, and it has no data and 'loading' is true. And then it could change to false again, but not without a glitch in UI or data loss.

To avoid re-renders and loss of data with Apollo hooks, hooks.ts and ProfileContext.ts use simple loadStateTracker that tracks when last update happened and did it happen after loading was actually requested.

If load status changes without an implicit load request, it is ignored. Component's data is updated only if last requested update happended after tracker's timestamp. It is a small hack, but when any component may alter centralized data, it must be controlled and kept safe.

Added new await-handler package for easier error handling with async/await