exposeProfileContext added to componentMocking
React hooks are exposed to tests with @testing-library/react-hooks. It can return all props returned from the hook.

useProfileContext required Apollo client so hook is wrapped with MockApolloClientProvider. And ProfileProvider to provide to hook to children.

waitForUpdate and waitForDataChange are helper functions for tests to wait for updates from graphQL queries. testing-library has waitForNextUpdate, but it does not update everytime.