ComponentMocking test improvements
Added BrowserRouter and ToastProvider because needed in Profile and ViewProfile tests

TestTools have helper functions waitforEl, waitForIsComplete and waitForDataChange so it is convenient to wait for data loading and re-rendering in tests

ProfileContextAsHTML component renders the profile context  as HTML and makes it easy to read updates. Hooks are not exposed like in exposeProfileContext and better to use HTML as source with Components than make work-arounds for hooks.