(HP-318) Add hook for mutations
The main purpose of this commit is "useProfileMutations.ts". It sends profile data updates to the server and updates cache after that. It uses ProfileContext and also passes through its data, so "useProfileDataEditor" is required to use only useProfileMutations and not also useProfileContext.

Testing mutations require a lot of new testing tools:
- "myProfileMocking.ts" has "cloneAndManipulateProfile" that provides functions for easy profile editing. It outputs new profile and also formValues needed to mock updates. New profile data is then used in mock responses and formValues when creating update variables.
- "componentMocking.ts" exposes the "useProfileMutations" hook for testing. Added also simple "createResultPropertyTracker" that creates a promise based listener for hook's properties.  "ProfileContextFetcher" component just auto-fetches profile data for the hook.

Also bundled all clean-up functions for tests to "cleanComponentMocks" function and updated ProfileContext to use it like the new UseProfileMutations.tests.ts

useProfileDataEditor hook now uses useProfileMutations instead of useProfileContext