(HP-727) Replace deprecated Dropdown
Select-component is used instead of Dropdown. If option list is long (countries in addresses) Combobox is used, because it has search functions. Side-effect of search is possibility to unselect. So added error handling to countries.

It is a known bug that Combobox error message  has different icon than textboxes.

Added label back to 'Asiontikieli' because it improves accessibility. Element is hidden with css, but aria-label uses the label prop as a prefix.

getA11yStatusMessage reads the selected item as "Suomi valittu" or "Maa ei ole valittuna".