(HP-879) Added authService.getAuthenticatedUser
If an authenticated user refreshes the page, session is valid on page load, but oidc client events are not triggered. The 'userLoaded' event is triggered only when renew is complete and it useless in this case.

React component should not start the poller, so added authService.getAuthenticatedUser which replaces the first authService.getUser() call. Polling is started when this function is called and user is already logged in.

Also it was better to move the isAuthenticatedUser check from Profile.tsx to authService and let it reject the call if user object is not valid.