(HP-962) Created a controller for cookie consents
The controller handles cookie read/write actions and stores consents until user saves them to cookie.

Created cookie is accessible from *.hel.fi and *.hel.ninja (and any *.domain.suffix).

Common conset 'names' are declared so same consents can be used between services and and asked again.

Cookie's "secure" -property is set to 'false' so it is readable and settable without https and therefore in localhost. This is not considered as a security risk because of the content of the cookie and 'sameSite' is set to 'strict' so the cookie is not sent to third party domains.

Added tests and required mocks.