(HP-1012) Use common css in Login.tsx
Removed css styles that are inherited from common css or conflict them.

The unusual ".content.content" is used to override common css.

Cannot set the order where css files are added to HTML. The specificity of the rule is doubled with ".class.class" which is better than using "!important".

Login page is vertically centered so top and bottom padding have same values.