(HP-1012) Use common css in AcccessibilityStatement.tsx
Use common naming with class names: "wrapper" and "content" instead of "page-wrapper" and "content-wrapper"

PageLayout also added to this view and removed header and pageMeta

AccessibilityStatement was the last page to use white bg. Now that all pages use same bg color, the footer should not overflow to the content area.

Added also "exact" to the route in App.tsx