(HP-1057) Added DeleteProfile tests
Added dataTestIds to ExpandingPanel or elements could not be found in DOM

Added a mocked Service Connections query result and a way to add it to mocked responses

Only user interactions and load states of the "service connection" -query are tested. Actual deletion is in useDeleteProfile-hook.

generateAllowedDataFieldEdge() was added to fix a code smell.