(HP-1099) Render the verify email notification only with certain amr
The notification should only be shown if user's amr is 'heltunnistussuomifi'.

In other cases the backend won't send an email and the notification should not be shown.

Because useProfile is mocked AND imported, the tunnistusSuomifiAMR cannot be reused inside this test.

Typescript compiler will fail. That's why 'tunnistusSuomifi' from AmrStatic is used and 'tunnistusSuomifiAMR' is re-declared.