(HP-1099) Force amr to be an array
Tunnistamo returns amr as a string, but OIDC spec states it is an array of strings

Amr is now converted to an array if it is a string. Undefined or empty amr is converted to an empty array

Removed the AMR type as it was useless. typeof config.helsinkiAccountAM is "any", so the AMR type is then also "any".