(HP-1169) Pass language to all ServiceConnections queries
Added a helper for query variables: createServiceConnectionsQueryVariables

Added also query variable typings to the useQuery/useLazyQuery

Placed all types to typings.ts. It is not possible to export imported interfaces directly.
That's why "ServiceConnectionsQueryVariables as ServiceConnectionsQueryVars" was added.

Detecting language change became obsolete as data is refetched when vars change.