(HP-1048) Schema and form data change
Old "value"-property is split in to "countryCallingCcode" and "number"

Number can be only numbers and min length of 5. Max length changed from 255 to 250 because country calling code is max 5 chars.

countryCallingCode is a required field. When creating a profile countryCallingCodes is only required if number is set.

Combined phone schemas in to getDefaultPhoneSchema(). They only differ in required()