(HP-1048) FormikDropdown update
Virtualized prop added to improve combobox performance.

There is an accessibility caveat: number of items is not resolvable (from HDS docs).

There is also a testing caveat: virtualized prop wont work with Jest (reported to HDS)

NODE_ENV is checked because of that.

The component could not handle language changes. The text in input field did not change.

Had to store selected option in state to fix it. And store last language too.

When language changes the label of the selected option is checked.

If old and new label differ, the selected option is changed.

If currentOption is used, then the form is handled from parent component. No need to check language.

Current option allows reverting the option for example in Additionalnformation when saving fails

Added missing tests because the component is now complicated.

Fixed some usages of the FormikDropdown.