(HP-1337) Added history tracker
Focus is set on page load only, if user is not on first page. Which means history has changed.

But redirections have to be ignored. User is redirected for example when not authorized and goes to Profile page.

History is tracked until internal routing event without redirection is detected. Then tracking is stopped.

'history' package was not in the package.json which raised code smells.

Added it with matching version to react-router dependency. Yarn lock did not update.