(HP-1337) Added a focus setter hook and focusable H1 component
Focus setter uses history tracker and sets focus when hasInternalPageLoads is true.

It's a bit odd to pass "disableFocusing" as a prop, but hooks cannot be used conditionally.

Focus setting is not wanted everytime.

Focus is set to first found element with given selector or certain class name. Fallback is a h1 element.

Added a component for focusable H1

Focusable element needs certain element attributes which this component sets.