(HP-1336) Fixed expading content button text for screen readers
The button text had a "clarifying" text which indicated, if the content is open or not.

That text conflicted with screen readers ability to detect expanding content.

The text read aloud was about "Show information <Content title> , closed" or "Hide information <Content title>, opened".

Resolution is to use just "<Content title>, closed/opened"