(HP-1425) Store the type of a node and restore it to saved data
Value of EmailType, AddressType and PhoneType was always reset to "OTHER. Default node properties where always copied to the saved data.

Types are not editable and were not stored, so those were overridden every time.

Now the type is stored to editData ja set to the save data.

Changed values of some node types in mock data and fixed + added tests.