(HP-1526) Change bg colors and component alignment
White on white was removed. The content area gets a litte wider as box padding is gone.

ProfileInformation bottom area needs to be grey all the way to the sides.

Removed user name with koros

To fill full width of the page, the content padding must be moved inside components. And components must be wrapped.

So ViewProfile has no css or extra elements anymore and those are moved to ServiceConnections and ProfileInformation.

Fixed tests. User name in "view-profile-heading" was removed and caused failures. Added new testIds.