(HP-1526) Remove white boxes from Profile page
White on white was removed. The content are gets a litte wider as box padding is gone.

ProfileInformation bottom area needs to be grey all the way to the sides.

To fill full width of the page, the content padding needs to move inside components. And components must be wrapped.

So ViewProfile has no css or extra elements anymore and those are moved to ServiceConnections and ProfileInformation.