(HP-1657) Added a iframe url loader
When getting an authorization code the browser jumps to Tunnistamo and back.

Same can be achieved with a iframe when the code if fetched like an ajax call.

Ajax cannot be used, because Tunnistamo requires its own cookies where the session is stored.

The loader creates an iframe and loads an url and waits for a window.postMessage.

When message is received it is validated and when message is correct one, the promise returned by the loader is resolved.

The loader timeouts in 30 sec.

Multiple iframes can co-exist.  The loader can be aborted with a window.postMessage