(HP-1657) Update graphql queries
Removed GdprServiceConnectionsQuery and use ServiceConnectionsQuery for both.

When the profile is deleted, ServiceConnections are fetched and then a dryrun is made to check if deletion is possible.

Authorization code requires service.gdprQueryScopes and UI service.title. Instead of 2 queries, one will do both.

Also makes the code base cleaner.

Added a mutation for deleting a single service connection