(HP-1686) Show failed / successful service connections in delete profile results
If deletion is not successful, the mutation returns which services fail.

Show lists in UI.

Updated translations.

Removed graphQL error parsing in DeleteProfileError

When GdprDeleteMyProfileMutation is set to return results object, the mutation can only return error when mutation failure is not "CONNECTED_SERVICE_DELETION_NOT_ALLOWED_ERROR".

So it is now useless to check error.graphQlErrors.

If there is an error, it is "fatal" and generic error message should be displayed.

Fixed tests. Also test in DeleteProfile named "When service connection load fails". It did not test service connection loading.