(HP-962) Added cookie consent modal
Modal is not shown when path is the auto login path as it is only a redirect path.

Modal is not shown on cookies page either. The user won't usually land there first and the content is identical with cookie page.

Modal and page have their own context (content) and do not affect each other. So making changes in one, do not change the other.

For that reason too they should not be shown at the same time.

Re-arranged language and route providers. Language provider rarely re-renders and should therefore be outside of route provider.