(HP-962) Helper for enabling / disabling tracking
One place to control Matomo permissions. If Matomo cookies are not accepted, all tracking is disabled. Currently there are no guidelines should anonymous tracking be enabled if user does not accept cookies. Tracking can  be enabled later, when a decision is made.

According to Matomo docs, it can track users without cookies and it is compliant with GDPR-rules.

https://matomo.org/faq/new-to-piwik/how-do-i-use-matomo-analytics-without-consent-or-cookie-banner/

https://matomo.org/faq/new-to-piwik/how-can-i-still-track-a-visitor-without-cookies-even-if-they-decline-the-cookie-consent/

When user does not give consent for cookies, Matomo does not add identifying payload to tracking events sent to Matomo or create cookies.

Matomo also clears set cookies, if consent is revoked.

This can be verified by checking cookies and requests in browsers inspector.

If cookies are allowed, application tab in the inspector shows cookies. Those are removed, when consent is revoked.

Requests do not have "_id" parameter in url when cookies are not allowed.