(HP-1734) Added a hook for tracking authenticated user activity
If user is not doing anything for a certain period of time, the user's tokens should not be renewed. Session should then die at the BE.

Inactive users get a warning first and if no activity is still not detected, user is logged out.

Tokens of an inactive user won't be renewed. Expiring tokens are renewed, if user becomes active again.

ActivityTracker component uses the hook and renders a warning, when duration of the inactivity is going to exceed the logout threshold.