(HP-1796) Move editing tools to a hook from multiItemEditor
There can be only one address / phone number. That has been changed before.

The components for both dataTypes were very similar before, but now it is better to remove the multiItem editor which handles both.

Components are now more unique and harder to keep them logically in one, so first move functionalities to a hook and then split the component.

The hook is easier to re-use later.