(HP-1796) Added new tests for AddressEditor
The editor should not save and post unedited values. Same functionality was already in emailEditor.

Also a new test for a bug when saving same data for a second time after an error.

Second attempt was not done, because data was not changed. Save attempt resulted in save success message.