(HP-1796) Harmonize editing result notifications
There was few different styles. Now one for success and one for error notifications

Errors are always shown above the buttons, so they won't possibly be placed below the fold - especially on mobile.

Success messages are shown on the bottom of the component.

Notifications are shown in many places and some of those require different spacings so added variants as props to the component.