(login-to) Get user syncronously from store
UserManager.getUser is async. Just to support all sync/async stores. This implementation uses a sessionStorage which is syncronous.

Async getUser makes usage harder  no reason. The user can be fetched manually, because the key used to store the user is kind of fixed.

Tests can easily check the key is correct.