(HP-2023) Added action getters and options
Options.idleWhenActive is for cases when an action starts a browser redirection and is then complete.

Then next action will wait until browser returns, but the action.active will be set to "true", so queue is not progressing to next action.

When browser returns and getNext() is called to proceed, then the action with active:true and idleWhenActive is picked as next.

Action getters filter actions and returns matching action(s) or undefined.