(HP-2023) Added new action option "syncronousCompletion"
Sometimes an action must be completed immediately, so the queue can resume later from next action.

For example when the browser is redirected, the action is closed, so the action is stored as complete.

Then when the browser returns, the queue will be resumed from next action.

Browser may stop execution when browser is redirected, so pending promise may not resolve.