(HP-2023) Action.options and action.data
Options.idleWhenActive is for cases when an action starts a browser redirection and is then complete.

Then next action will wait until browser returns, but the action.active will be set to "true", so queue is not progressing to next action.

When browser returns and getNext() is called to proceed, then the action with active:true and idleWhenActive is picked as next.

Options.noStorage indicates the action results should not be stored to session storage. Used with large data sets like graphQL queries.

Options.syncronousCompletion is used when an action must be completed immediately, so the queue can resume later from next action.

For example when the browser is redirected, the action is completed immediately, so the action is stored as complete.

Then when the browser returns, the queue will be resumed from next action.

Options.resumable indicates the queue can resume from the action. Queue is resumed when browser redirections are complete.

Action.data can hold any kind of data the action needs. Like name of the service to delete.