(HP-2023) Action for creating auth code request props
This action creates state and oidcUri parameters for the auth code requests

State is stored and compared when browser returns from the oidc server.

There is separate action for tunnistamo and keycloak.

Gdpr scopes are checked and if either oidc server auth code is not needed, it is not fetched.