(HP-2023) Action for redirecting to the oidc provider
This action creates proper urls for getting auth codes and redirects the browser

There is separate action for tunnistamo and keycloak.

Gdpr scopes are checked and if either oidc server auth code is not needed, that action resolves immediately.

Redirection is slightly delayed, so the promise is resolved for sure. Some browsers may halt the js excution when browser url changes.