(HP-2023) Internal redirection actions
When auth codes are ready, the browser must be redirected back to the page the process started from.

Start page is download/delete profile page or service connections page.

The redirection is done from the gdpr-callback page when auth codes are ready.

There is an action for redirecting ("redirector" action) back to the start page.

And action for detecting ("catcher" action) when redirection is done.

The start page is stored to the action.data.

If the browser was just redirected with "window.location.href" the whole page would be reloaded. This is not best way to redirect  a single page app. History should be used.

HIstory works only in hooks, which is not possible to use in actions.

So the redirection data is stored to the action's result/errorMessage.

The redirection data will be read from there when an action is complete.

Redirections also hold the failed action type or the action type which should catch the redirection.

Nothing is reading the results atm.