(HP-2023) Hook for using the queues
The hook gets the required queue and restores the one in the session storage.

The hook determines next phase to execute and controls the queue. It exposes helper functions and the state of the queue.

Added mocks for the queues so they can be tested one by one and each action can be triggered manually.

All queues use almost the same actions, so there is no need  to test every queue in detail.

Each action have their own tests anyway.