(hp-2126) Fix race condition with useEffect
An action with idleWhenActive:true can be resumed, when resume() is called in useEffect. This throws an error.

This can happen when an action is executed, but useEffect calls resume() before next render and nextPhase is still 'resumeWithAuthCodes'.

getActionStatus() now checks the action for idleWhenActive and is there a pending  promise.