(hp-2126) Handle url redirection delays
When the user has no service connections, the browser is not really redirected.

The redirector action sets new url, but the redirectionCatcher is activated immediately, but url changes later.

When browser is actually redirected the redirectionCatcher is executed the same way, but again when url has changed and components have mounted again.